requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mutagen==1.46.0
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
drf-yasg==1.21.6
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
packaging==23.1
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
twilio==8.5.0
uritemplate==4.1.1
urllib3==2.0.4
yarl==1.9.2


